+++ gulpfile.js
@@ -128,7 +128,7 @@ gulp.task("doc:modules",
 gulp.task("build:modules", function (done) {
 
     runSequence(
-        "lint:modules",
+        //"lint:modules",
         "clean:dist",
         "copy:modules:npm",
         "copy:modules:dist",