+++ modules/core/src/model/dynamic-date-control.model.ts
@@ -4,12 +4,14 @@ import { serializable } from "../decorator/serializable.decorator";
 
 export interface DynamicDateControlModelConfig extends DynamicFormValueControlModelConfig<Date> {
 
+    format?: string;
     max?: Date;
     min?: Date;
 }
 
 export abstract class DynamicDateControlModel extends DynamicFormValueControlModel<Date> {
 
+    @serializable() format: string | null;
     @serializable() max: Date | null;
     @serializable() min: Date | null;
 
@@ -17,6 +19,7 @@ export abstract class DynamicDateControlModel extends DynamicFormValueControlMod
 
         super(config, cls);
 
+        this.format = config.format || null;
         this.max = config.max || null;
         this.min = config.min || null;
     }